user_auth/health/happ/requirements.txt
django==5.0.3
kubernetes==28.1.0
djangorestframework==3.15.1
